backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.36
pymysql==1.1.1
python-jose==3.3.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
pydantic>=2.0
pydantic-settings>=2.0
openpyxl==3.1.5
python-slugify==8.0.1
